WebAppDeployment/code/requirements.txt
# Essential dependencies only
langchain>=0.1.0
langchain-ollama>=0.1.0
pyyaml>=6.0
requests>=2.31.0
